+++ tidy3d/components/simulation.py
@@ -420,9 +420,17 @@ class AbstractYeeGridSimulation(AbstractSimulation, ABC):
     def _validate_boundary_spec_symmetry(cls, val, values):
         """Error if symmetry is imposed along an axis but the boundary conditions are not the same
         on both sides."""
+
+        def equivalent(plus: BoundarySpec, minus: BoundarySpec) -> bool:
+            """Returns whether two boundary conditions are physically identical."""
+            # Make copies of `plus` and `minus` with the `name` attribute set to "".
+            plus_cpy = plus.updated_copy(name="")
+            minus_cpy = minus.updated_copy(name="")
+            return plus_cpy == minus_cpy
+
         boundaries = [val.x, val.y, val.z]
         for ax, symmetry, ax_bounds in zip("xyz", values.get("symmetry"), boundaries):
-            if symmetry != 0 and ax_bounds.plus != ax_bounds.minus:
+            if symmetry != 0 and not equivalent(ax_bounds.plus, ax_bounds.minus):
                 raise ValidationError(
                     f"Symmetry '{symmetry}' along axis {ax} requires the same boundary "
                     f"condition on both sides of the axis."