+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Giving opposite boundaries different names no longer causes a symmetry validator failure.
+
 ## [2.9.0rc2] - 2025-07-17
 
 ### Added